Day 6
The first day I got stuck. The reason is that I had made a tiny mistake
in the movement. If it hits the obstacle, I made it turn and move ahead
one step in one loop, instead, I should just make it turn, and the
movement will go to the next loop.
The reason is that if there is 2 obstacles like this
--->*...
...*....
turning and moving is wrong because it will directly move to another
obstacle without triggering.